+++ src/main/java/services/product/ExploreProductsServlet.java
@@ -1,4 +1,4 @@
-package services.products;
+package services.product;
 
 import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
@@ -6,11 +6,12 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.SneakyThrows;
 
-@WebServlet("/user/list_products")
-public class ListProductsServlet extends HttpServlet {
+@WebServlet("/user/explore_products")
+public class ExploreProductsServlet extends HttpServlet {
+
     @Override
     @SneakyThrows
-    protected void doPost(HttpServletRequest request, HttpServletResponse response) {
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp) {
 
     }
 }
